Split out globals.typedarray.d.ts
