+++ types/node/index.d.ts
@@ -22,7 +22,7 @@
  * IN THE SOFTWARE.
  */
 
-// NOTE: These definitions support NodeJS and TypeScript 4.9+.
+// NOTE: These definitions support NodeJS and TypeScript 5.7+.
 
 // Reference required types from the default lib:
 /// <reference lib="es2020" />
@@ -30,12 +30,15 @@
 /// <reference lib="esnext.intl" />
 /// <reference lib="esnext.bigint" />
 
+// Definitions specific to TypeScript 4.9 through 5.7+
+/// <reference path="globals.typedarray.d.ts" />
+/// <reference path="buffer.d.ts" />
+
 // Base definitions for all NodeJS modules that are not specific to any version of TypeScript:
 /// <reference path="assert.d.ts" />
 /// <reference path="assert/strict.d.ts" />
 /// <reference path="globals.d.ts" />
 /// <reference path="async_hooks.d.ts" />
-/// <reference path="buffer.d.ts" />
 /// <reference path="child_process.d.ts" />
 /// <reference path="cluster.d.ts" />
 /// <reference path="console.d.ts" />
