+++ types/node/v16/ts5.6/index.d.ts
@@ -31,12 +31,13 @@
 /// <reference lib="esnext.bigint" />
 
 // Definitions specific to TypeScript 4.9 through 5.6
-/// <reference path="./globals.d.ts" />
+/// <reference path="./globals.typedarray.d.ts" />
 /// <reference path="./buffer.d.ts" />
 
 // Base definitions for all NodeJS modules that are not specific to any version of TypeScript:
 /// <reference path="../assert.d.ts" />
 /// <reference path="../assert/strict.d.ts" />
+/// <reference path="../globals.d.ts" />
 /// <reference path="../async_hooks.d.ts" />
 /// <reference path="../child_process.d.ts" />
 /// <reference path="../cluster.d.ts" />
