+++ types/node/v20/globals.d.ts
@@ -400,20 +400,6 @@ declare global {
             unref(): this;
         }
 
-        type TypedArray<TArrayBuffer extends ArrayBufferLike = ArrayBufferLike> =
-            | Uint8Array<TArrayBuffer>
-            | Uint8ClampedArray<TArrayBuffer>
-            | Uint16Array<TArrayBuffer>
-            | Uint32Array<TArrayBuffer>
-            | Int8Array<TArrayBuffer>
-            | Int16Array<TArrayBuffer>
-            | Int32Array<TArrayBuffer>
-            | BigUint64Array<TArrayBuffer>
-            | BigInt64Array<TArrayBuffer>
-            | Float32Array<TArrayBuffer>
-            | Float64Array<TArrayBuffer>;
-        type ArrayBufferView<TArrayBuffer extends ArrayBufferLike = ArrayBufferLike> = TypedArray<TArrayBuffer> | DataView<TArrayBuffer>;
-
         interface Require {
             (id: string): any;
             resolve: RequireResolve;
